Add 'backend:reload_routes' task to reload Router's routes
Sometimes Router API and Router can get out of sync with one
another, e.g. immediately after a data sync. This should correct
itself as soon as a route is added, edited or removed, as Router
API then gets Router to reload all routes. However, if no route
changes, it is useful to be able to force a Router reload via
manual means.

We intend to call this Rake task post data sync in govuk-puppet.